@@ -1,16 +1,12 @@
 import React from 'react';
-import { BrowserRouter as Router } from 'react-router-dom';
-
 import Routes from './routes';
 
 import GlobalStyle from './styles/global';
 
-const App: React.FC = () => (
+const App = () => (
   <>
     <GlobalStyle />
-    <Router>
-      <Routes />
-    </Router>
+    <Routes />
   </>
 );
 
