@@ -1,15 +1,13 @@
 import React from 'react';
-
 import { FiPlusSquare } from 'react-icons/fi';
 import { Container } from './styles';
 
 import Logo from '../../assets/logo.svg';
-
 interface IHeaderProps {
   openModal: () => void;
 }
 
-const Header: React.FC<IHeaderProps> = ({ openModal }) => (
+const Header = ({ openModal }: IHeaderProps) => (
   <Container>
     <header>
       <img src={Logo} alt="GoRestaurant" />
